requirements.txt
streamlit
sentence-transformers
pinecone-client
pdfplumber
python-dotenv
google-genai
